implemented mapping of local prints to ams trays
I implemented logging of the mqtt messages to replay them and not to print everytime I need to test something.
For this I seperated some code into functions.
The logging needs some work, so that the filesize doesn't grow to much.

To map the trays to the filament we need to know after how many changes which filament is used first.
So I extract the gcode file and parse it for tray changes.
When we know this, we can track the tray changes via mqtt and assign the filaments.
After all filaments have been assigned we can spend the spools.